requirements.txt
docx2pdf==0.1.8
docx_mailmerge==0.5.0
pandas==2.1.0
pdfplumber==0.10.2
pytest==7.4.1
python_docx==0.8.11
pywin32==306
rich==13.5.2
textual==0.36.0
